+++ src/Mods/AstralArena/ScriptExtender/Lua/AstralArena/Shared/Constants.lua
@@ -1,10 +1,16 @@
 return {
     ModuleUUID = "29c48c80-8777-f7b5-6bb8-376c1c5d8db6",
-    DisplayVersion = "0.3.2-alpha.3",
+    DisplayVersion = "0.3.2-alpha.4",
     ArenaLevel = "AA_Arena_Main",
     SchemaVersion = 1,
     EntrantCount = 8,
     MaxPartySize = 4,
+    MenuMessages = {
+        defeat = "h7d1ea950dfe44604a4c32268059678c1",
+        draw = "hf930087eaa3246688664f96e14c64c64",
+        reward = "h0621de70c9aa4efba27b86b471d34682",
+        recipient = "h91755f48cd8d485b8d15e756b1e224d7",
+    },
     Levels = {
         quarterfinal = 5,
         semifinal = 8,